splitted base definition ready to convert it to ABC, remove all the unused engines
